+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to semantic versioning for configuration changes.
 
 ## [Unreleased]
 
+### Auto-updated - 2026-01-18 09:00:17
+- Modified: Traccar-Binding/src/main/java/org/openhab/binding/traccar/internal/TraccarDeviceHandler.java
+- Modified: items/traccar.items
+
+## [Unreleased]
+
 ### Auto-updated - 2026-01-18 08:45:06
 - Modified: Traccar-Binding/src/main/java/org/openhab/binding/traccar/internal/TraccarDeviceHandler.java
 - Modified: Traccar-Binding/src/main/resources/OH-INF/thing/thing-types.xml